requirements.txt
albumentations==1.3.1
deep_sort_realtime==1.3.2
fairscale==0.4.13
ipython==8.15.0
matplotlib==3.8.2
moviepy==1.0.3
numpy==1.23.0
onnxruntime==1.13.1
opencv_python==4.8.1.78
opencv_python_headless==4.8.0.74
pandas==2.1.4
Pillow==10.1.0
Pillow==10.1.0
psutil==5.9.0
psutil==5.9.6
pycocotools==2.0.6
pytube==15.0.0
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
super_gradients==3.5.0
torch==2.1.1
torchinfo==1.8.0
torchmetrics==0.8.0
torchvision==0.16.1
tqdm==4.66.1
ultralytics==8.0.223
vision_transformers==0.1.1.0
yt_dlp==2023.11.16
